Increase uniquness of the BDBid
BDBid is the index which uniquely identify each client (even different browsers on same device).
It's based on the second the first sync is made multiplied by 1000.
There is a (slight) chance of conflict. To increase the uniqueness we add at creation a random number.
NOTE: looks ok but still need FULL test!